refactor: core engine remove duplicated hashing func
