+++ README.md
@@ -55,7 +55,6 @@ You can use both to have different values on client-side and server-side.
 
 ---
 
-
 # Usage
 
 ## A). 🌟 **App directory** (new)
