--- lib/src/core/client/strategy/util.ts
@@ -1,5 +1,0 @@
-import * as murmurHash3 from 'murmurhash3js';
-
-export default function normalizedValue(id: string, groupId: string, normalizer = 100): number {
-  return (murmurHash3.x86.hash32(`${groupId}:${id}`) % normalizer) + 1;
-}
